[NFC] Reorganize lifetime dependence utilities
