+++ include/swift/AST/LifetimeDependence.h
@@ -17,12 +17,18 @@
 #ifndef SWIFT_AST_LIFETIMEDEPENDENCE_H
 #define SWIFT_AST_LIFETIMEDEPENDENCE_H
 
-#include "swift/AST/Decl.h"
+#include "swift/AST/Identifier.h"
+#include "swift/AST/IndexSubset.h"
 #include "swift/Basic/Debug.h"
 #include "swift/Basic/OptionSet.h"
-#include "llvm/Support/TrailingObjects.h"
+#include "swift/Basic/SourceLoc.h"
+
+#include "llvm/ADT/ArrayRef.h"
 
 namespace swift {
+
+class AbstractFunctionDecl;
+class LifetimeDependentReturnTypeRepr;
 
 enum class LifetimeDependenceKind : uint8_t {
   Copy = 0,
@@ -121,6 +127,42 @@ public:
         "Invalid LifetimeDependenceSpecifier::LifetimeDependenceKind");
   }
 };
+
+class LifetimeDependenceInfo {
+  IndexSubset *inheritLifetimeParamIndices;
+  IndexSubset *borrowLifetimeParamIndices;
+  IndexSubset *mutateLifetimeParamIndices;
+
+  static llvm::Optional<LifetimeDependenceInfo>
+  fromTypeRepr(AbstractFunctionDecl *afd, Type resultType, bool allowIndex);
+
+public:
+  LifetimeDependenceInfo()
+      : inheritLifetimeParamIndices(nullptr),
+        borrowLifetimeParamIndices(nullptr),
+        mutateLifetimeParamIndices(nullptr) {}
+  LifetimeDependenceInfo(IndexSubset *inheritLifetimeParamIndices,
+                         IndexSubset *borrowLifetimeParamIndices,
+                         IndexSubset *mutateLifetimeParamIndices)
+      : inheritLifetimeParamIndices(inheritLifetimeParamIndices),
+        borrowLifetimeParamIndices(borrowLifetimeParamIndices),
+        mutateLifetimeParamIndices(mutateLifetimeParamIndices) {}
+
+  operator bool() const { return !empty(); }
+
+  bool empty() const {
+    return inheritLifetimeParamIndices == nullptr &&
+           borrowLifetimeParamIndices == nullptr &&
+           mutateLifetimeParamIndices == nullptr;
+  }
+
+  std::string getString() const;
+  void Profile(llvm::FoldingSetNodeID &ID) const;
+
+  static llvm::Optional<LifetimeDependenceInfo>
+  get(AbstractFunctionDecl *decl, Type resultType, bool allowIndex = false);
+};
+
 } // namespace swift
 
 #endif
