+++ lib/SIL/IR/TypeLowering.cpp
@@ -3682,7 +3682,7 @@ getFunctionInterfaceTypeWithCaptures(TypeConverter &TC,
           .withConcurrent(funcType->isSendable())
           .withAsync(funcType->isAsync())
           .withIsolation(funcType->getIsolation())
-		  .withLifetimeDependenceInfo(funcType->getLifetimeDependenceInfo())
+          .withLifetimeDependenceInfo(funcType->getLifetimeDependenceInfo())
           .build();
 
   return CanAnyFunctionType::get(
