+++ lib/Sema/TypeCheckDecl.cpp
@@ -2576,128 +2576,6 @@ static Type validateParameterType(ParamDecl *decl) {
   return Ty;
 }
 
-llvm::Optional<LifetimeDependenceInfo> validateLifetimeDependenceInfo(
-    LifetimeDependentReturnTypeRepr *lifetimeDependentRepr, Type resultTy,
-    Decl *decl, bool allowIndex) {
-  auto *afd = cast<AbstractFunctionDecl>(decl);
-  auto *dc = decl->getDeclContext();
-  auto &ctx = dc->getASTContext();
-  auto &diags = ctx.Diags;
-  SmallBitVector inheritLifetimeParamIndices(afd->getParameters()->size() + 1);
-  SmallBitVector borrowLifetimeParamIndices(afd->getParameters()->size() + 1);
-  SmallBitVector mutateLifetimeParamIndices(afd->getParameters()->size() + 1);
-
-  auto updateLifetimeDependenceInfo = [&](LifetimeDependenceSpecifier specifier,
-                                          unsigned paramIndexToSet,
-                                          ValueOwnership ownership) {
-    auto loc = specifier.getLoc();
-    auto kind = specifier.getLifetimeDependenceKind();
-
-    if (resultTy->isEscapable()) {
-      diags.diagnose(loc, diag::lifetime_dependence_invalid_return_type);
-      return true;
-    }
-
-    if (ownership == ValueOwnership::Default) {
-      diags.diagnose(loc, diag::lifetime_dependence_missing_ownership_modifier);
-      return true;
-    }
-    if (kind == LifetimeDependenceKind::Borrow &&
-        ownership != ValueOwnership::Shared) {
-      diags.diagnose(loc, diag::lifetime_dependence_cannot_use_kind, "borrow",
-                     getOwnershipSpelling(ownership));
-      return true;
-    }
-    if (kind == LifetimeDependenceKind::Mutate &&
-        ownership != ValueOwnership::InOut) {
-      diags.diagnose(loc, diag::lifetime_dependence_cannot_use_kind, "mutate",
-                     getOwnershipSpelling(ownership));
-      return true;
-    }
-    if (kind == LifetimeDependenceKind::Consume &&
-        ownership != ValueOwnership::Owned) {
-      diags.diagnose(loc, diag::lifetime_dependence_cannot_use_kind, "consume",
-                     getOwnershipSpelling(ownership));
-      return true;
-    }
-    if (inheritLifetimeParamIndices.test(paramIndexToSet) ||
-        borrowLifetimeParamIndices.test(paramIndexToSet)) {
-      diags.diagnose(loc, diag::lifetime_dependence_duplicate_param_id);
-      return true;
-    }
-    if (kind == LifetimeDependenceKind::Copy ||
-        kind == LifetimeDependenceKind::Consume) {
-      inheritLifetimeParamIndices.set(paramIndexToSet);
-    } else if (kind == LifetimeDependenceKind::Borrow) {
-      borrowLifetimeParamIndices.set(paramIndexToSet);
-    } else {
-      assert(kind == LifetimeDependenceKind::Mutate);
-      mutateLifetimeParamIndices.set(paramIndexToSet);
-    }
-    return false;
-  };
-
-  for (auto specifier : lifetimeDependentRepr->getLifetimeDependencies()) {
-    switch (specifier.getSpecifierKind()) {
-    case LifetimeDependenceSpecifier::SpecifierKind::Named: {
-      bool foundParamName = false;
-      unsigned paramIndexToSet = 1;
-      for (auto *param : *afd->getParameters()) {
-        if (param->getParameterName() == specifier.getName()) {
-          foundParamName = true;
-          if (updateLifetimeDependenceInfo(specifier, paramIndexToSet,
-                                           param->getValueOwnership())) {
-            return llvm::None;
-          }
-          break;
-        }
-        paramIndexToSet++;
-      }
-      if (!foundParamName) {
-        diags.diagnose(specifier.getLoc(),
-                       diag::lifetime_dependence_invalid_param_name,
-                       specifier.getName());
-        return llvm::None;
-      }
-      break;
-    }
-    case LifetimeDependenceSpecifier::SpecifierKind::Ordered: {
-      auto paramIndex = specifier.getIndex();
-      if (paramIndex > afd->getParameters()->size()) {
-        diags.diagnose(specifier.getLoc(),
-                       diag::lifetime_dependence_invalid_param_index,
-                       paramIndex);
-        return llvm::None;
-      }
-      if (updateLifetimeDependenceInfo(
-              specifier, /*paramIndexToSet*/ specifier.getIndex() + 1,
-              afd->getParameters()->get(paramIndex)->getValueOwnership())) {
-        return llvm::None;
-      }
-      break;
-    }
-    case LifetimeDependenceSpecifier::SpecifierKind::Self: {
-      if (!afd->hasImplicitSelfDecl()) {
-        diags.diagnose(specifier.getLoc(),
-                       diag::lifetime_dependence_invalid_self);
-        return llvm::None;
-      }
-      if (updateLifetimeDependenceInfo(
-              specifier, /*selfIndex*/ 0,
-              afd->getImplicitSelfDecl()->getValueOwnership())) {
-        return llvm::None;
-      }
-      break;
-    }
-    }
-  }
-
-  return LifetimeDependenceInfo(
-      IndexSubset::get(ctx, inheritLifetimeParamIndices),
-      IndexSubset::get(ctx, borrowLifetimeParamIndices),
-      IndexSubset::get(ctx, mutateLifetimeParamIndices));
-}
-
 static void maybeAddParameterIsolation(AnyFunctionType::ExtInfoBuilder &infoBuilder,
                                        ArrayRef<AnyFunctionType::Param> params) {
   if (hasIsolatedParameter(params))
@@ -2872,15 +2750,7 @@ InterfaceTypeRequest::evaluate(Evaluator &eval, ValueDecl *D) const {
       resultTy = TupleType::getEmpty(AFD->getASTContext());
     }
 
-    auto *returnTypeRepr = AFD->getResultTypeRepr();
     llvm::Optional<LifetimeDependenceInfo> lifetimeDependenceInfo;
-    if (returnTypeRepr) {
-      if (auto *lifetimeDependentRepr =
-              dyn_cast<LifetimeDependentReturnTypeRepr>(returnTypeRepr)) {
-        lifetimeDependenceInfo = validateLifetimeDependenceInfo(
-            lifetimeDependentRepr, resultTy, D, /*allowIndex*/ false);
-      }
-    }
 
     // (Args...) -> Result
     Type funcTy;
@@ -2897,9 +2767,13 @@ InterfaceTypeRequest::evaluate(Evaluator &eval, ValueDecl *D) const {
       // Defer bodies must not escape.
       if (auto fd = dyn_cast<FuncDecl>(D))
         infoBuilder = infoBuilder.withNoEscape(fd->isDeferBody());
-      if (lifetimeDependenceInfo.has_value())
+
+      lifetimeDependenceInfo = LifetimeDependenceInfo::get(AFD, resultTy);
+      if (lifetimeDependenceInfo.has_value()) {
         infoBuilder =
             infoBuilder.withLifetimeDependenceInfo(*lifetimeDependenceInfo);
+      }
+
       auto info = infoBuilder.build();
 
       if (sig && !hasSelf) {
@@ -2915,9 +2789,10 @@ InterfaceTypeRequest::evaluate(Evaluator &eval, ValueDecl *D) const {
       auto selfParam = computeSelfParam(AFD);
       AnyFunctionType::ExtInfoBuilder selfInfoBuilder;
       maybeAddParameterIsolation(selfInfoBuilder, {selfParam});
-      if (lifetimeDependenceInfo.has_value())
+      if (lifetimeDependenceInfo.has_value()) {
         selfInfoBuilder =
             selfInfoBuilder.withLifetimeDependenceInfo(*lifetimeDependenceInfo);
+      }
       // FIXME: Verify ExtInfo state is correct, not working by accident.
       auto selfInfo = selfInfoBuilder.build();
       if (sig) {
